Move resizable to base
This is a generally useful component, should be in base
