+++ src/vs/base/browser/ui/resizable/resizable.ts
